add tcpdump support for upload Mode on Server
In this path, we introduce the support for running tcpdump on the server
when the client requires the upload mode.  This enhancement allows for
efficient network traffic monitoring, specifically tailored for upload
operations.

Users can now specify the network interface that tcpdump will attach to
by using the --interface parameter directly from the Server command
line. This feature is intended to improve the server's ability to
capture and analyze network packets in real-time, providing valuable
insights during upload processes.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>